fix: bf16 issue

+++ training/train_deepspeed.py
@@ -500,10 +500,7 @@ def train(args):
             for k, v in adjacency_init.items()
         }
         engine.module.visual.initialize_backbones(
-            adjacency_init_device,
-            in_channels=channels_hint,
-            device=engine.device,
-            dtype=torch.bfloat16,
+            adjacency_init_device, in_channels=channels_hint, device=engine.device
         )
 
         if dist.is_initialized():
